Enabling backing up extra folders (#30)
This adds functionality to backup extra folders from the container into their own extras zip file associated with the container. It includes behaviors to trim the extras included in the zip file as well.

```
java:
    - name: <container name>
      extras:
        - <folder path>
      worlds:
        - <world path>
```

The result is a series of zip files. So if I have a container called “minecraft_server” and the world was “Yosemite”, the result would be:

minecraft_server.extras.Timestamp.zip
Yosemite.Timestamp.zip