Add Basic HTTP Request Handling on 8080 (#101)
Provides an implementation of `/health` and `/live` for monitoring service health. It also provides a partial implementation of `/status` although it doesn't report sizes of backups properly yet, just provides when the last backup occurred, if it was successful, and which containers were backed up. 

It also implements `/start-backup` which requires the caller provide the correct token. The token is generated randomly at the start of the service, and written to a file. The file path can be configured using `tokenPath` in the configuration file, but by default it creates a `.bedrockifierToken` file adjacent to the configuration file, similar to `.service_is_healthy` 

To kick off a backup, a user with access to the configuration directory, and curl can do something like the following. 

```
curl http://{IPADDRESS}:8080/start-backup
   -H "Authorization: Bearer $(cat configDir/.bedrockifierToken)"
```